Update caja.h
Float a caja

+++ caja.h
@@ -11,24 +11,24 @@ class Caja
 {
 private:
     std::string nombre;
-    int xMin;
-    int yMin;
-    int zMin;
-    int xMax;
-    int yMax;
-    int zMax;
+    float xMin;
+    float yMin;
+    float zMin;
+    float xMax;
+    float yMax;
+    float zMax;
     std::list<Objeto> l_objetos;
     std::list<Vertice> l_vertices;
     std::list<Cara> l_caras;
 
 public:
     Caja();
-    void setxMin(int xMin);
-    void setyMin(int yMin);
-    void setzMin(int zMin);
-    void setxMax(int xMax);
-    void setyMax(int yMax);
-    void setzMax(int zMax);
+    void setxMin(float xMin);
+    void setyMin(float yMin);
+    void setzMin(float zMin);
+    void setxMax(float xMax);
+    void setyMax(float yMax);
+    void setzMax(float zMax);
     int getxMin();
     int getyMin();
     int getzMin();
